Update backend tests
